Add SourceController (#3602)
